Change nproc to long to match sysconf() return type

+++ samu.c
@@ -203,7 +203,7 @@ main(int argc, char *argv[])
 argdone:
 	if (!buildopts.maxjobs) {
 #ifdef _SC_NPROCESSORS_ONLN
-		int nproc = sysconf(_SC_NPROCESSORS_ONLN);
+		long nproc = sysconf(_SC_NPROCESSORS_ONLN);
 		switch (nproc) {
 		case -1: case 0: case 1:
 			buildopts.maxjobs = 2;
